refactor: LottoStatistics.java WinningLotto.java 책임 분리 1. LottoStatistics.java 수익률을 통계하는 클래스로 정의 2. WinningLotto 당첨번호와 보너스볼을 포장하고 당첨번호를 match하는 클래스로 정의
